refactor: split game data implementation
